% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_for_onestep.R
\name{estimateCensoring}
\alias{estimateCensoring}
\title{Estimation for the Method of Cause-Specific Hazards}
\usage{
estimateCensoring(dataList, adjustVars, t0, SL.ctime = NULL,
  returnModels = FALSE, verbose = TRUE, gtol = 0.001, env, testdat,
  ...)
}
\description{
Estimation for the Method of Cause-Specific Hazards
}
